Moved ConsoleWriter to svc/cli
