--- go/svc/cli/result/writer.go
@@ -1,4 +1,0 @@
-package result
-
-type Writer interface {
-}
